requirements.txt
streamlit
pandas
psycopg2-binary
pytz
streamlit-extras
bcrypt
pytest